Cleanup things from testing
This commit fixes several issues identified by testing the early
prototypes. The first thing is abandon the use of GitPython, while in
theory a pythonic interface to interacting with git would be easier in
practice it was more difficult to use then just calling out to git. Then
some other small typos and oversights are fixed and it looks like asv is
being called correctly now using test webhook requests.